password.c up to FUN_801a2c78

+++ include/password.h
@@ -5,7 +5,7 @@
 #include "num_level_36260.h"
 
 void PS1_EncryptPassword(void);
-u8 PS1_VerifyDecryptPassword(void);
+s32 PS1_VerifyDecryptPassword(void);
 void FUN_801a17c8(u8 arg0);
 u8 PS1_GetLevelFromPassword(void);
 void PS1_GeneratePassword_LivesCount(u8 lives_count);
@@ -16,7 +16,7 @@ u8 PS1_ValidatePassword(void);
 u8 PS1_GeneratePassword(void);
 void PS1_LoadSaveFromPassword(void);
 u8 PS1_AttemptLoadSaveFromPassword(void);
-void PS1_UnusedGenerateAndPrintPassword(s16 param_1, s16 param_2, u8 param_3, u32 param_4);
+void PS1_UnusedGenerateAndPrintPassword(s16 param_1, s16 param_2, u8 param_3, u8 param_4);
 void PS1_ClearPassword(void);
 void FUN_801a2c78(void);
 void FUN_801a2d40(void);
